SolveROMs -> Interpolate/Rollouts
SolveROMs was always a bit of a strange file with strange functions. Average_roms and sample_roms were strange names for the functions. Since both methods essentially solve the LD for a testing parameter combination given by Interpolate, I felt it was best to move them into the interpolate sub-directory (in a new "Rollouts.py" file). I renamed "average_rom" and "sample_rom" to "Mean_Rollout" and "Sample_Rollouts", since these names better align with what the functions do.

I also moved the "Generate_Heatmap_Data" file into the Plot file, where it belongs.

This allowed me to delete the old, strange SolveROMs.py file. I updated the readme and other files to reflect these changes.
